Add support to stop a running game on a channel
