+++ .sqlx/query-169fcfa537d37ed33d6f77f6a8661c1d54cd00248d285105a73dc6bf27e38f20.json
@@ -1,0 +1,16 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "\n            UPDATE \n                channel\n            SET \n                hour = $1,\n                active = $2\n            WHERE\n                id = $3;\n            ",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Bool",
+        "Text"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "169fcfa537d37ed33d6f77f6a8661c1d54cd00248d285105a73dc6bf27e38f20"
+}
